requirements.txt
Pillow
keyboard
sounddevice
numpy
